PR feedback fixes
