--- example-queries/bulk-download-query.graphql
@@ -1,29 +1,0 @@
-query BulkDownloadsQuery{
-      bulkDownloads{
-          id
-          status
-          startedAt
-          ownerUserId
-          analysisCount
-          entityInputFileType
-          sampleNames
-          rawInputsJson {
-            description
-            downloadType
-            downloadDisplayName
-            fileFormat
-          }
-          file {
-            downloadLink {
-              url
-            }
-            size
-          }
-          toDelete{
-            user_name
-            log_url
-            totalSamples
-            progress
-          }
-        }
-    }`;
