--- example-queries/bulk-download-with-limit-query.graphql
@@ -1,37 +1,0 @@
-query BulkDownloadsWithLimitQuery(
-    $limit: Int, 
-    $searchBy:String
-  ) {
-      bulkDownloads(
-        input: {
-          searchBy: $searchBy, 
-          limit: $limit
-        }
-      ) {
-          id
-          status
-          startedAt
-          ownerUserId
-          analysisCount
-          entityInputFileType
-          sampleNames
-          rawInputsJson {
-            description
-            downloadType
-            downloadDisplayName
-            fileFormat
-          }
-          file {
-            downloadLink {
-              url
-            }
-            size
-          }
-          toDelete{
-            user_name
-            log_url
-            totalSamples
-            progress
-          }
-        }
-    }`;
